Bug Fix for multi client connection
Multiple clients were unable to connect because the server changed its initial dir once 1 client had connected to C:\Users\...\CWD, preventing access to the json file. The json file needs to be in the initial dir so it isn't be deleted by clients, so the code now resets the dir on connect. 
NOTE - There is still a bug on new account creation. For some reason creating a new account bugs out the whole thing and I'm not sure why.