Bug Fix for protocol.py
The Upload and Download portion of file_transfer would only work for txt files. It is now working for all file types.
NOTE: There is currently a bug related to the download/upload rate where if it goes two quick the code will think their is a divide by 0 error. It is handled properly, and does not imped file transfer, but it of note.