Another bug fix fir protocol.py
While file_transfer was able to handle all file types, it was sending all the information at once, meaning it could not handle very large files. It now separates it into chunks to handle larger files. In theory it should be able to handle up to 2 GB (as described by the assignment), but I don't have a file of that size to test it (I know it can handle up to 150 MB)